+++ packages/strapi-admin/admin/src/containers/SettingsPage/index.js
@@ -5,20 +5,15 @@
  */
 
 import React from 'react';
-import { Route, Switch, useRouteMatch } from 'react-router-dom';
 import { LeftMenu, LeftMenuList } from 'strapi-helper-plugin';
 
 import Webhooks from '../Webhooks';
 import Wrapper from './Wrapper';
 
 function SettingsPage() {
-  let { path } = useRouteMatch();
-  const renderWebhooks = props => <Webhooks {...props} />;
-
   const menuItems = [
     {
-      name: 'General',
-      searchable: false,
+      title: { id: 'app.components.LeftMenuLinkContainer.general' },
       links: [
         {
           title: 'webhooks',
@@ -36,15 +31,13 @@ function SettingsPage() {
           <div className="col-md-3">
             <LeftMenu>
               {menuItems.map(item => {
-                return <LeftMenuList {...item} key={item.name} />;
+                return <LeftMenuList {...item} key={item.title} />;
               })}
             </LeftMenu>
           </div>
           <div className="col-md-9">
             {/* TODO when needed - Routing */}
-            <Switch>
-              <Route path={path} render={renderWebhooks} />
-            </Switch>
+            <Webhooks />
           </div>
         </div>
       </div>
