--- packages/strapi-admin/admin/src/containers/Webhooks/EditView.js
@@ -1,17 +1,0 @@
-/**
- *
- * EditView
- *
- */
-
-import React from 'react';
-
-function EditView() {
-  return (
-    <div>
-      <p>Edit</p>
-    </div>
-  );
-}
-
-export default EditView;
