--- packages/strapi-admin/admin/src/containers/Webhooks/Wrapper.js
@@ -1,15 +1,0 @@
-/**
- *
- * Wrapper
- *
- */
-
-import styled from 'styled-components';
-
-const Wrapper = styled.div`
-  > div:first-of-type {
-    margin-bottom: 33px;
-  }
-`;
-
-export default Wrapper;
