+++ packages/strapi-admin/admin/src/plugins.js
@@ -27,10 +27,10 @@ module.exports = {
     .default,
   'strapi-plugin-content-type-builder': require('../../../strapi-plugin-content-type-builder/admin/src')
     .default,
-  // 'strapi-plugin-documentation': require('../../../strapi-plugin-documentation/admin/src')
-  //   .default,
-  // 'strapi-plugin-email': require('../../../strapi-plugin-email/admin/src')
-  //   .default,
+  'strapi-plugin-documentation': require('../../../strapi-plugin-documentation/admin/src')
+    .default,
+  'strapi-plugin-email': require('../../../strapi-plugin-email/admin/src')
+    .default,
   'strapi-plugin-upload': require('../../../strapi-plugin-upload/admin/src')
     .default,
 };
