+++ packages/strapi-admin/admin/src/translations/en.json
@@ -221,8 +221,8 @@
   "Settings.webhook.list.description": "Get POST changes notifications.",
   "Settings.webhook.list.button.add": "Add new webhook",
   "Settings.webhook.list.button.delete": "Delete",
-  "Settings.webhook.list.label.sing": "webhook",
-  "Settings.webhook.list.label.plur": "webhooks",
+  "Settings.webhook.list.label.singular": "webhook",
+  "Settings.webhook.list.label.plurarl": "webhooks",
   "app.containers.App.notification.error.init": "An error occured while requesting the API",
   "components.Input.error.password.noMatch": "Passwords do not match",
   "form.button.done": "Done",
