+++ packages/strapi-helper-plugin/lib/src/components/LeftMenuHeader/Search.js
